Organize UI modules into subpackage
