+++ main.py
@@ -9,7 +9,7 @@ from PyQt5.QtWidgets import QApplication, QSplashScreen
 from PyQt5.QtGui import QIcon, QPixmap, QFont
 from PyQt5.QtCore import Qt, QTimer
 
-from src.menu_window import MenuWindow
+from src.ui.menu_window import MenuWindow
 from local_activation.activacion import run_activation
 
 # Toggle for enabling or disabling the license check. Set to ``True`` to
