+++ src/ui/memoria_window.py
